refactor: split renderer.py into focused modules
_subpaths.py  — bezier subpath geometry (_subpath_to_3n1, _classify_subpaths, _SubpathChild)
_state_keys.py — hashable state cache keys (_VTKey, _MathTexKey, _ImgKey, _VGroupKey, _VMobKey)
_camera.py    — camera helpers (_needs_camera_loop, _serialize_camera)

renderer.py shrinks from 1389 → 1133 lines.
widget.py and tests updated to import from canonical locations.
